✅ challenge: open file to line number
It is possible to open zipped files using vim.
`vim <file>.zip`.
You can move through the files and editing any of the files is persistent.

You can also open files from a name. Let's say you are in a file
that references another file name. Put the cursor at the beginning of
the file name, then click `gf` which stands for 'goto file'.

If the file name references a specific line such as "<file>:<line>".
Then you can use `g<SHIFT>f` to go to the specific line in that file.