✅ challenge: external command
You can run external commands and paste their output into the
current file using `:r !<command>`.

It is also possible to alter the contents of the current file using
an external command such as using jq (a JSON formatter).
You do it as follows:
`% !<command> .`
Basically (%) use the current file, (!) run external command <command>
and put the output back into this file (.).
For the jq example it would be: `% !jq .`.